ErrorBudget.jl Documentation

Error budget computation.

Table of contents

ErrorBudget.jl Documentation

Table of contents

Introduction

Package Features

Introduction

julia> using ErrorBudget

Package Features

Tank accuracy computation

Function documentation

ErrorBudget.aiming_error — Method

aiming_error(aiming)

Returns the alignment error (horizontal and vertical) for the error sources defined in 'aiming'

ErrorBudget.alignment_error — Method

alignment_error(align)

Returns the alignment error (horizontal and vertical) for the error sources defined in 'align'

ErrorBudget.azimuthError — Method

azimuthError(σ, target,proj,weapon,tank,aero;args )

Returns the normal distribution for the impact points (vertical and horizontal distribution). The function performs a Monte Carlo simulation by sampling the azimuth angle. It is assumes that the elevation angle variation is a gaussian distribution with μ = the computed azimuth angle and standard deviation = σ:

tank

weapon : the launcher of type Gun

proj : the projectile

target

aero : the aerodynamic coefficients of the projectile

One optional argument is N (the number of Monte Carlo simulations). Default = 100

ErrorBudget.cantError — Method

cantError(σ, σmeas, target,proj,weapon,tank,aero, atmosphere;args )

Returns the normal distribution for the impact points (vertical and horizontal distribution). The function performs a Monte Carlo simulation by sampling the cant angle. It is assumes that the cant variation is a gaussian distribution with μ = tank.hull.Φ and standard deviation = σ. On this value of cant we perform another sampling to take into account the measurement error. We assume also a normal distribution with μ = sampled cant and the standard deviation = σmeas:

tank

weapon : the launcher of type Gun

proj : the projectile

target

aero : the aerodynamic coefficients of the projectile

One optional argument is N (the number of Monte Carlo simulations). Default = 100

ErrorBudget.createTargetCirc — Method

createTargetCirc(size, position)

Returns a circular target by specifying the radius (size) and the position (range)

ErrorBudget.createTargetRect — Method

createTargetRect(a,b, position)

Return a rectangular target by specifying the length 'a', the height 'b' and the 'position' (range)

ErrorBudget.elevationError — Method

elevationError(σ, target,proj,weapon,tank,aero;args )

Returns the normal distribution for the impact points (vertical and horizontal distribution). The function performs a Monte Carlo simulation by sampling the elevation angle. It is assumes that the elevation angle variation is a gaussian distribution with μ = the compute elevation angle and standard deviation = σ:

tank

weapon : the launcher of type Gun

proj : the projectile

target

aero : the aerodynamic coefficients of the projectile

One optional argument is N (the number of Monte Carlo simulations). Default = 100

ErrorBudget.flight_error — Method

flight_error(flight)

Returns the projectile dispersion (horizontal and vertical)

ErrorBudget.interTable — Method

interTable(range,jumpTable, value)

Returns the 'value' from an interpolation table 'jumpTable' over the 'range'

ErrorBudget.lof_error — Method

lof_error(lof)

Returns the line of fire error (horizontal and vertical) for the error sources defined in 'lof'

ErrorBudget.muzzleVelError — Method

muzzleVelError(σ, target,proj,weapon,tank,aero;args )

Returns the normal distribution for the impact points (vertical and horizontal distribution). The function performs a Monte Carlo simulation by sampling the muzzle velocity. It is assumes that the muzzle velocity variation is a gaussian distribution with μ = weapon.u₀ and standard deviation = σ:

tank

weapon : the launcher of type Gun

proj : the projectile

target

aero : the aerodynamic coefficients of the projectile

One optional argument is N (the number of Monte Carlo simulations). Default = 100

ErrorBudget.rangeError — Method

rangeError(target, σ, tank, weapon, proj, aero; args)

Returns the normal distribution for elevation and azimuth angles. The function performs a Monte Carlo simulation by sampling the target range. It is assumes that the range variation is a gaussian distribution with μ = target.ρ and standard deviation = σ. For each sampled range, a elevation and azimuth angle is computed:

tank

weapon : the launcher of type Gun

proj : the projectile

target

aero : the aerodynamic coefficients of the projectile

One optional argument is N (the number of Monte Carlo simulations). Default = 100

ErrorBudget.temperatureError — Method

temperatureError(σ, target,proj,weapon,tank,aero, atmosphere;args )

Returns the normal distribution for the impact points (vertical and horizontal distribution). The function performs a Monte Carlo simulation by sampling the air temperature. It is assumes that the temperature variation is a gaussian distribution with μ = atmosphere.t and standard deviation = σ:

tank

weapon : the launcher of type Gun

proj : the projectile

target

aero : the aerodynamic coefficients of the projectile

atmosphere: the atmospheric characteristics

One optional argument is N (the number of Monte Carlo simulations). Default = 100

ErrorBudget.variableBias — Method

variableBias(target,tank, weapon, proj, aero, wind, atmosphere;args ...)

Returns the variable bias error as an objet "Error". optional arguments are all the standard deviation of error sources: σrange, σcrosswind,σrangewind, σjump,σfc,σzeroing,σboresight,σMv, σtemp, σpress, σopt, σgun, σcant, σcant_meas default values for optional parameters are nothing or nul:

tank

weapon : the launcher of type Gun

proj : the projectile

target

aero : the aerodynamic coefficients of the projectile

wind

atmosphere

ErrorBudget.windCrossError — Method

windCrossError(σ,wind,tank, weapon,proj, target, aero;args ...)

Returns the normal distribution for the impact points (horizontal coordinate). The function performs a Monte Carlo simulation by sampling the wind (cross field). It is assumes that the cross wind variation is a gaussian distribution with μ = wind.cross and standard deviation = σ:

tank

weapon : the launcher of type Gun

proj : the projectile

target

aero : the aerodynamic coefficients of the projectile

One optional argument is N (the number of Monte Carlo simulations). Default = 100

ErrorBudget.windRangeError — Method

windRangeError(σ,wind,tank, weapon,proj, target, aero;args ...)

Returns the normal distribution for the impact points (vertical coordinate). The function performs a Monte Carlo simulation by sampling the wind (range field). It is assumes that the range wind variation is a gaussian distribution with μ = wind.range and standard deviation = σ:

tank

weapon : the launcher of type Gun

proj : the projectile

target

aero : the aerodynamic coefficients of the projectile

One optional argument is N (the number of Monte Carlo simulations). Default = 100

ErrorBudget.σ — Method

σ(aiming,align,lof,flight)

Returns the total error (horizontal and vertical) with error budget defined by error sources:

aiming: aiming error

align: alignment error

lof: line of fire error

flight: projectile dispersion

Type documentation

ErrorBudget.AimingError — Type

AimingError(point_def,human,res,opt_path)

defines an objet 'AimingError' with fields containing the aiming error sources:

point_def: error (standard deviatin) on the aimpoint definition expressed as a tupple (horizontal and vertical).

human: human error (standard deviatin) on the aimpoint definition expressed as a tupple (horizontal and vertical).

res: resolution error (standard deviatin) on the aimpoint definition expressed as a tupple (horizontal and vertical).

opt_path: optical path bending error (standard deviatin) on the aimpoint definition expressed as a tupple (horizontal and vertical).

ErrorBudget.AlignmentError — Type

AlignmentError(boresight_acc,boresight_ret,sync,stab, gun_pos)

defines an objet 'AlignmentError' with fields containing the alignment error sources:

boresight_acc: boresight accuracy (standard deviatin) expressed as a tupple (horizontal and vertical).

boresight_ret: boresight retention (standard deviatin) expressed as a tupple (horizontal and vertical).

sync: synchronism/parallelism error (standard deviatin) expressed as a tupple (horizontal and vertical).

stab: stabilisation error (standard deviation) expressed as a tupple (horizontal and vertical).

gun_pos: gun position measurement error (standard deviation) expressed as a tupple (horizontal and vertical).

ErrorBudget.Error — Type

Error(horizontal, vertical)

Returns the Error object with field: 'horizontal' and 'vertical'

ErrorBudget.FlightError — Type

FlightError(aero)

defines an objet 'FlightError' which contains the projectile dispersion as an error (standard deviation) expressed as a tupple (horizontal and vertical).

ErrorBudget.LofError — Type

LofError(fcs,lead,cant,earth_rate, muzzle_vel, temp, press, cross_wind, range_wind, range, jump, zeroing)

defines an objet 'LofError' with fields containing the line of fire error sources:

fcs: fire control error (standard deviation) expressed as a tupple (horizontal and vertical).

lead: lead angle (standard deviation) expressed as a tupple (horizontal and vertical).

cant: cant angle error (standard deviation) expressed as a tupple (horizontal and vertical).

earth_rate: earth rate error (standard deviation) expressed as a tupple (horizontal and vertical).

muzzle_vel: muzzle velocity error (standard deviation) expressed as a tupple (horizontal and vertical).

temp: temperature error (standard deviation) expressed as a tupple (horizontal and vertical).

press: pressure error (standard deviation) expressed as a tupple (horizontal and vertical).

cross_wind: cross wind error (standard deviation) expressed as a tupple (horizontal and vertical).

range_wind: range wind error (standard deviation) expressed as a tupple (horizontal and vertical).

jump: fleet jump error (standard deviation) expressed as a tupple (horizontal and vertical).

zeroing: calibration error (standard deviation) expressed as a tupple (horizontal and vertical).

ErrorBudget.SpheError — Type

SpheError(σx,σy,σz,μx,μy,μz)

Returns a spherical error object with standard deviation and center of impact points

ErrorBudget.UnitEffect — Type

UnitEffect(velocity,temp,density, windR,windC)

Returns the UnitEffects for different ranges. The parameters considered are: * velocity * temp: temperature * density * windR: wind range * windC: cross wind

ErrorBudget.AimingError

ErrorBudget.AlignmentError

ErrorBudget.Error

ErrorBudget.FlightError

ErrorBudget.LofError

ErrorBudget.SpheError

ErrorBudget.UnitEffect

ErrorBudget.aiming_error

ErrorBudget.alignment_error

ErrorBudget.azimuthError

ErrorBudget.cantError

ErrorBudget.createTargetCirc

ErrorBudget.createTargetRect

ErrorBudget.elevationError

ErrorBudget.flight_error

ErrorBudget.interTable

ErrorBudget.lof_error

ErrorBudget.muzzleVelError

ErrorBudget.rangeError

ErrorBudget.temperatureError

ErrorBudget.variableBias

ErrorBudget.windCrossError

ErrorBudget.windRangeError

ErrorBudget.σ